fix eslint, refactor action footer to specific files
