+++ src/utils/componentUtils/index.js
@@ -1,4 +1,3 @@
-export * from './actionFooter';
 export * from './choosePlayerModal';
 export * from './cityLines';
 export * from './board';
@@ -6,3 +5,6 @@ export * from './cityMarkers';
 export * from './playerMarkers';
 export * from './playerMenu';
 export * from './chooseCardModal';
+export * from './build';
+export * from './move';
+export * from './share';
